asdf

+++ java_project/DAM1 24-25 BRD/src/ud5/clasesejercicios/E080123/Hora.java
@@ -1,19 +1,19 @@
 package ud5.clasesejercicios.E080123;
 
 public class Hora {
-    protected int hora;
-    protected int minuto;
+    protected byte hora;
+    protected byte minuto;
 
     // Constructor
     public Hora(int hora, int minuto) {
         if (hora >= 0 && hora < 24) {
-            this.hora = hora;
+            this.hora = (byte) hora;
         } else {
             this.hora = 0;
         }
 
         if (minuto >= 0 && minuto < 60) {
-            this.minuto = minuto;
+            this.minuto = (byte) minuto;
         } else {
             this.minuto = 0;
         }
@@ -34,7 +34,7 @@ public class Hora {
     // Asigna un nuevo valor a los minutos si es válido
     public boolean setMinuto(int valor) {
         if (valor >= 0 && valor < 60) {
-            this.minuto = valor;
+            this.minuto = (byte) valor;
             return true;
         }
         return false;
@@ -43,7 +43,7 @@ public class Hora {
     // Asigna un nuevo valor a la hora si es válido
     public boolean setHora(int valor) {
         if (valor >= 0 && valor < 24) {
-            this.hora = valor;
+            this.hora = (byte) valor;
             return true;
         }
         return false;
